requirements.txt
python-telegram-bot
requests
terminaltables
beautifulsoup4
